# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Car Rental API implementation.",
        "title": "Car Rental API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "eibrahimarisoy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cars/": {
            "get": {
                "description": "List all cars with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "car"
                ],
                "summary": "List all cars",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pickup date",
                        "name": "pickup_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pickup time",
                        "name": "pickup_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dropoff date",
                        "name": "dropoff_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dropoff time",
                        "name": "dropoff_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID formatted ID",
                        "name": "location_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.CarListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a car with payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "car"
                ],
                "summary": "Create a car",
                "parameters": [
                    {
                        "description": "Car payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.CarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.CarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/locations/": {
            "get": {
                "description": "List all locations with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "List all locations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.Pagination"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a location with payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Create a location",
                "parameters": [
                    {
                        "description": "Location payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/location.LocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/location.LocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/offices/": {
            "get": {
                "description": "List all offices with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "offices"
                ],
                "summary": "List all offices",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.Pagination"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a office with payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "offices"
                ],
                "summary": "Create a office",
                "parameters": [
                    {
                        "description": "Office payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/office.OfficeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/office.OfficeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/_type.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "_type.APIErrorResponse": {
            "type": "object",
            "properties": {
                "errCode": {
                    "type": "integer"
                },
                "errDetails": {},
                "errMessage": {
                    "type": "string"
                }
            }
        },
        "car.CarListResponse": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/car.CarResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "q": {
                    "type": "string"
                },
                "rows": {},
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "car.CarRequest": {
            "type": "object",
            "properties": {
                "fuel": {
                    "type": "string",
                    "enum": [
                        "gas",
                        "diesel",
                        "electric"
                    ]
                },
                "name": {
                    "type": "string"
                },
                "office_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "available",
                        "rented"
                    ]
                },
                "transmission": {
                    "type": "string",
                    "enum": [
                        "automatic",
                        "manual"
                    ]
                },
                "vendor_id": {
                    "type": "string"
                }
            }
        },
        "car.CarResponse": {
            "type": "object",
            "properties": {
                "fuel": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office": {
                    "$ref": "#/definitions/office.OfficeResponse"
                },
                "status": {
                    "type": "string"
                },
                "transmission": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/vendors.VendorResponse"
                }
            }
        },
...
        "location.LocationRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "location.LocationResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "office.OfficeRequest": {
            "type": "object",
            "properties": {
                "closing_hours": {
                    "type": "string"
                },
                "location_id": {
                    "type": "string"
                },
                "opening_hours": {
                    "type": "string"
                },
                "vendor_id": {
                    "type": "string"
                },
                "working_days": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "office.OfficeResponse": {
            "type": "object",
            "properties": {
                "closing_hours": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/location.LocationResponse"
                },
                "opening_hours": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/vendors.VendorResponse"
                },
                "working_days": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pagination.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "q": {
                    "type": "string"
                },
                "rows": {},
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "vendors.VendorResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}